# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mappa API",
    "description": "Planning and hazard assistant for Puerto Rico. Everything a frontend needs is under /api/v1; see docs/API.md.",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/tiles/{name}/{z}/{x}/{y}.mvt": {
      "get": {
        "tags": [
          "tiles"
        ],
        "summary": "Tile",
        "description": "One vector tile, built from PostGIS on demand.",
        "operationId": "tile_api_v1_tiles__name___z___x___y__mvt_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tiles/{name}.json": {
      "get": {
        "tags": [
          "tiles"
        ],
        "summary": "Tilejson",
        "description": "The layer as a vector source - pass this URL straight to a map library.",
        "operationId": "tilejson_api_v1_tiles__name__json_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TileJSON"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TileJSON": {
        "properties": {
          "tilejson": {
            "type": "string",
            "title": "Tilejson",
            "default": "3.0.0"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tiles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tiles"
          },
          "minzoom": {
            "type": "integer",
            "title": "Minzoom"
          },
          "maxzoom": {
            "type": "integer",
            "title": "Maxzoom"
          },
          "bounds": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Bounds"
          },
          "vector_layers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Vector Layers"
          }
        },
        "type": "object",
        "required": [
          "name",
          "tiles",
          "minzoom",
          "maxzoom",
          "bounds",
          "vector_layers"
        ],
        "title": "TileJSON",
        "description": "A TileJSON 3.0 descriptor - accepted directly by MapLibre GL and Mapbox GL."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}